delete object-store only iff pre-existing

+++ main.js
@@ -22,7 +22,9 @@ let vectorLayer = new VectorLayer();
 
 const idb = await openDB('hofapp', 3, {
   upgrade(db, oldVersion, newVersion, transaction, event) {
-    db.deleteObjectStore('imgs');
+    if(db.objectStore(imgs)) {
+      db.deleteObjectStore('imgs');
+    }
     const imgStore = db.createObjectStore('imgs',  { keyPath: 'url' });
     imgStore.createIndex("url", "url");
   }
